# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartFlow Next",
    "version": "0.1.0"
  },
  "paths": {
    "/api/session": {
      "get": {
        "summary": "Session Info",
        "operationId": "session_info_api_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "METHOD_NOT_ALLOWED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-required-permission": "session:read"
      }
    },
...
    "/api/openapi.json": {
      "get": {
        "summary": "Openapi",
        "operationId": "openapi_api_openapi_json_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIDocument"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "METHOD_NOT_ALLOWED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-required-permission": "schema:read"
      }
    },
    "/api/errors": {
      "get": {
        "summary": "Errors",
        "operationId": "errors_api_errors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/ErrorInfo"
                  },
                  "type": "object",
                  "title": "Response Errors Api Errors Get"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "METHOD_NOT_ALLOWED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-required-permission": "schema:read"
      }
    },
    "/api/jobs": {
      "post": {
        "summary": "Create",
        "operationId": "create_api_jobs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 80,
              "pattern": "^[a-zA-Z0-9_-]+$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJob"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "METHOD_NOT_ALLOWED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Domain conflict; inspect error.code and recovery",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-permission": "jobs:create"
      },
      "get": {
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_REJECTED or PERMISSION_DENIED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "METHOD_NOT_ALLOWED",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "INPUT_INVALID; safe field locations only",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR; inspect trace",
            "headers": {
              "X-Trace-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-required-permission": "jobs:read"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthMode": {
        "type": "string",
        "enum": [
          "local_session",
          "dev_bypass"
        ],
        "title": "AuthMode"
      },
...
      "CreateJob": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "scenario": {
            "type": "string",
            "enum": [
              "success",
              "auth_required",
              "unknown_send",
              "save_failure",
              "transient",
              "pending"
            ],
            "title": "Scenario",
            "default": "success"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CreateJob"
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "recovery": {
            "type": "string",
            "title": "Recovery"
          },
          "http_status": {
            "type": "integer",
            "title": "Http Status"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "prepare",
                  "generate",
                  "save",
                  "done"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "recovery",
          "http_status",
          "trace_id",
          "stage"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/InvalidField"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "OpenAPIDocument": {
        "properties": {
          "openapi": {
            "type": "string",
            "title": "Openapi"
          },
          "info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Info"
          },
          "paths": {
            "additionalProperties": true,
            "type": "object",
            "title": "Paths"
          },
          "components": {
            "additionalProperties": true,
            "type": "object",
            "title": "Components"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "openapi",
          "info",
          "paths"
        ],
        "title": "OpenAPIDocument"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "browser_agent",
          "browser_pairing",
          "owner",
          "operator",
          "viewer",
          "support"
        ],
        "title": "Role"
      },
...
      "SessionResponse": {
        "properties": {
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "auth_mode": {
            "$ref": "#/components/schemas/AuthMode"
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "actor_id",
          "role",
          "auth_mode",
          "permissions"
        ],
        "title": "SessionResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}